+++ skia_demo/.gitignore
@@ -1,0 +1,5 @@
+*.beam
+*.ez
+/build
+erl_crash.dump
+/output
