+++ skia_demo/README.md
@@ -1,0 +1,24 @@
+# skia_demo
+
+[![Package Version](https://img.shields.io/hexpm/v/skia_demo)](https://hex.pm/packages/skia_demo)
+[![Hex Docs](https://img.shields.io/badge/hex-docs-ffaff3)](https://hexdocs.pm/skia_demo/)
+
+```sh
+gleam add skia_demo@1
+```
+```gleam
+import skia_demo
+
+pub fn main() {
+  // TODO: An example of the project in use
+}
+```
+
+Further documentation can be found at <https://hexdocs.pm/skia_demo>.
+
+## Development
+
+```sh
+gleam run   # Run the project
+gleam test  # Run the tests
+```
