+++ skia_demo/gleam.toml
@@ -1,0 +1,30 @@
+name = "skia_demo"
+version = "1.0.0"
+target = "javascript"
+
+
+[javascript]
+runtime = "deno"
+
+[javascript.deno]
+allow_ffi = true
+allow_env = true
+allow_read = true
+
+# Fill out these fields if you intend to generate HTML documentation or publish
+# your project to the Hex package manager.
+#
+# description = ""
+# licences = ["Apache-2.0"]
+# repository = { type = "github", user = "", repo = "" }
+# links = [{ title = "Website", href = "" }]
+#
+# For a full reference of all the available options, you can have a look at
+# https://gleam.run/writing-gleam/gleam-toml/.
+
+[dependencies]
+gleam_stdlib = ">= 0.44.0 and < 2.0.0"
+paint = { path = "../../paint" }
+
+[dev-dependencies]
+gleeunit = ">= 1.0.0 and < 2.0.0"
