+++ src/paint/event.gleam
@@ -24,6 +24,13 @@ pub type Event {
   MouseReleased(MouseButton)
 }
 
+pub type Next(state, output) {
+  /// Continue using the given state
+  Continue(state)
+  /// Terminate with some given output
+  Stop(output)
+}
+
 pub type Key {
   KeyLeftArrow
   KeyRightArrow
